refactor Logo component #694
