+++ WMIAdventure/frontend/src/js/components/global/atoms/Logo/index.js
@@ -1,0 +1,1 @@
+export {default} from './Logo';
