+++ WMIAdventure/frontend/src/js/components/global/atoms/Logo/styled-components/H1.js
@@ -1,11 +1,11 @@
 import styled from 'styled-components';
 
-const H1 = styled.div`
+const H1 = styled.h1`
   margin: 0;
   text-transform: uppercase;
   font-family: Hack, monospace;
   color: ${({theme}) => theme.colors.dark};
-  font-size: 20px;
+  font-size: ${({setFontSize}) => setFontSize ? setFontSize : '20px'};
   font-weight: ${({theme}) => theme.weight.bold};
 `;
