+++ WMIAdventure/frontend/src/js/components/global/atoms/Logo/styled-components/LogoImage.js
@@ -1,9 +1,9 @@
 import styled from 'styled-components';
 
-const Logo = styled.img`
+const LogoImage = styled.img`
   width: 40px;
   height: 40px;
   order: -1;
 `;
 
-export default Logo;
+export default LogoImage;
